==> 02a-hktvmall_DataCleaning_F&M.ipynb <==
import pandas as pd
import numpy as np
import string
import re
df=pd.read_csv("hktvmall_skincare_females_and_males.csv", index_col=0)

#"Discount Price" Cleaning
for i in range(0, len(df["Discount price"])):
    df["Discount price"][i]=df["Discount price"][i].replace(",","").strip()
    try: 
        df["Discount price"][i]=re.match("(\d+)\.?.*", df["Discount price"][i]).group(1)
    except:
        df["Discount price"][i]="0"

df["Discount price"]=df["Discount price"].astype("int64")

#"Original Price" Cleaning
for i in range(0, len(df["Original Price"])):
    
    df["Original Price"][i]=df["Original Price"][i].replace(",","").strip()

    try: 
        df["Original Price"][i]=re.match("(\d+)\.?.*", df["Original Price"][i]).group(1)
    except:
        df["Original Price"][i]="0"

df["Original Price"]=df["Original Price"].astype("int64")

#Create new features - "Discount" & "% Discount"
df["Discount"]=df["Original Price"]-df["Discount price"]
df["%Discount"]=df["Discount"]*(1/df["Original Price"])*100

#Create new feature - "Revenue"
df["Revenue"]=df["Discount price"]*df["Sales"]

#Rounding "Rating" & "Review"
df=df.round({'Rating': 1, 'Review=': 0})

#"Sales" Cleaning
for i in range(0, len(df["Sales"])):
    try:
        df["Sales"][i]=df["Sales"][i].replace(",","").strip()
    except:
        df["Sales"][i]=np.nan
df["Sales"]=df["Sales"].astype("float64")

#Create new feature - "Price Range"
df['Price Range'] = df.apply(lambda _: '', axis=1)
for i in range(0, len(df["Discount price"])):
    if df["Discount price"][i]<=100:
        df["Price Range"][i]="0-100"
    elif (df["Discount price"][i]>100) and (df["Discount price"][i]<=200):
        df["Price Range"][i]="101-200"
    elif (df["Discount price"][i]>200) and (df["Discount price"][i]<=300):
        df["Price Range"][i]="201-300"
    elif (df["Discount price"][i]>300) and (df["Discount price"][i]<=400):
        df["Price Range"][i]="301-400"
    elif (df["Discount price"][i]>400) and (df["Discount price"][i]<=500):
        df["Price Range"][i]="401-500"
    elif (df["Discount price"][i]>500) and (df["Discount price"][i]<=600):
        df["Price Range"][i]="501-600"
    elif (df["Discount price"][i]>600) and (df["Discount price"][i]<=700):
        df["Price Range"][i]="601-700"
    elif (df["Discount price"][i]>700) and (df["Discount price"][i]<=800):
        df["Price Range"][i]="701-800"
    elif (df["Discount price"][i]>800) and (df["Discount price"][i]<=900):
        df["Price Range"][i]="801-900"
    elif (df["Discount price"][i]>900) and (df["Discount price"][i]<=1000):
        df["Price Range"][i]="901-1000"
    elif df["Discount price"][i]>1000:
        df["Price Range"][i]=">1000"

#Create new feature - "Product Category"
df["Product Category"] = df.apply(lambda _: '', axis=1)
for i in range(0, len(df["Product Category Code"])):
    if df["Product Category Code"][i]=="AA16301000001":
        df["Product Category"][i]="Perfume for Women"

    elif df["Product Category Code"][i]=="AA16103010001":
        df["Product Category"][i]="Hydrating Mask"

    elif df["Product Category Code"][i]=="AA16202010001":
        df["Product Category"][i]="Lipstick"

    elif df["Product Category Code"][i]=="AA16101500001":
        df["Product Category"][i]="Toner"

    elif df["Product Category Code"][i]=="AA16102010101":
        df["Product Category"][i]="Serum"

    elif df["Product Category Code"][i]=="AA16102510101":
        df["Product Category"][i]="Moisture Cream"

    elif df["Product Category Code"][i]=="AA16402000001":
        df["Product Category"][i]="Gel Nail"

    elif df["Product Category Code"][i]=="AA16108000001":
        df["Product Category"][i]="Gifts/ Travel Set"

    elif df["Product Category Code"][i]=="AA16302000001":
        df["Product Category"][i]="Perfume for Men"

    elif df["Product Category Code"][i]=="AA16103510001":
        df["Product Category"][i]="Eye Cream"

    else:
        df["Product Category"][i]="Others"

#Save to .csv file
df.to_csv("hktvmall_skincare_females_and_males_cleaned.csv")
